Fix some error in RouterForComplexRoutes. (#3)
* Fix some error in RouterForComplexRoutes.

Hello,

First, congrats for this project. I love it!

This pull request fix two minors errors :
- If the controller method hasn't arguments, give it an empty array.
- If the container doesn't have the Controller service, fix the parameters route index to _route.

I also added an other check in order to see if the _controller parameter contains the correct separator (here "::"). This separator is now a constant to change it easily.

This helped me to be more explicit about the encounter \LogicException.

What do you think about it?

Thanks you for this little project

* Update RouterForComplexRoutes.php

Fix missing constant call